requirements.txt
numpy
pandas
xarray
cfgrib
eccodes
metpy
pytz
python-dotenv
langchain
langchain-community
langchain-openai
tqdm
argparse